requirements.txt
psycopg2-binary==2.9.9
confluent-kafka
flask==3.0.0
flask-cors==4.0.0
python-dateutil==2.8.2
pyspark
redis
